:recycle: [repositories] Move `nullprovider`, `constprovider`, `dictprovider` to `tests.fixtures`
